+++ tests/test_doc.py
@@ -47,8 +47,8 @@ class Child(lb.Device):
         
 
 class Mock(lb.Device):
-    ''' Helpful driver wrapper
-    '''
+    """ Helpful driver wrapper
+    """
     
     child0 = Child('addr0')
     child1 = Child('addr1')