+++ ChangeLog.md
@@ -6,10 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 **This is a significant change that is API incompatible with 0.20!** 
+
 ### Added
 - Unit tests for lb.concurrently and lb.sequentially in test_concurrently.py
 - `lb.NonScalar` data type
 - `lb.property` and `lb.method` decorators, which specify trait decorator behavior
+- `lb.hide_in_traceback` decorator scrubs the decorated function from tracebacks
 
 ### Changed
 - Show warnings on trait assignment typos like `device.frequency = 5` instead of `device.state.frequency = 5`